+++ flutter_fawry_pay_web/test/flutter_fawry_pay_web_test.dart
@@ -1,0 +1,6 @@
+import 'package:flutter/services.dart';
+import 'package:flutter_test/flutter_test.dart';
+import 'package:flutter_fawry_pay_web/flutter_fawry_pay.dart';
+
+void main() {
+}