+++ flutter_fawry_pay_platform_interface/CHANGELOG.md
@@ -1,3 +1,15 @@
+# 0.0.3
+
+- Ready for publish and use.
+
+# 0.0.2+1
+
+- Remove nullity.
+
+# 0.0.2
+
+- Add function randomAlphaNumeric.
+
 # 0.0.1
 
 - Initial open-source release.